fix: rename toggle component instead switch
